requirements.txt

flask
pyswisseph
